+++ csense-kotlin-common/src/main/kotlin/csense/kotlin/extensions/collections/map/MutableMap.kt
@@ -10,7 +10,8 @@ package csense.kotlin.extensions.collections.map
  *
  */
 inline fun <K, V> MutableMap<K, V>.removeAll(
-        crossinline predicate: Function1<Map.Entry<K, V>, Boolean>) {
+        crossinline predicate: Function1<Map.Entry<K, V>, Boolean>
+) {
     //find all to remove, and get the key
     val toRemove = entries.filter(predicate).mapKeys()
     toRemove.forEach { remove(it) }