+++ csense-kotlin-common/src/main/kotlin/csense/kotlin/extensions/collections/types.kt
@@ -1,5 +1,5 @@
 package csense.kotlin.extensions.collections
 
 
-typealias GenericGetterIndexMethod<T> = Function1<Int, T>
-typealias GenericSetterIndexMethod<T> = Function2<Int, T, Unit>
+typealias GenericGetterIndexMethod<T> = (index: Int) -> T
+typealias GenericSetterIndexMethod<T> = (index: Int, item: T) -> Unit